+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.7.0
 ## Added
+* Add /settings endpoint to view (GET) and reset (DELETE) config (modifications are not currently supported)
 
 ## Fixed